pyproject.toml
[project]
name = "echonote"
version = "0.1.0"
description = "A Django-powered personal blog site."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "django>=5.2.5,<6.0.0",
    "django-environ>=0.11",
    "Pillow>=10.0",
    "django-taggit>=5.0",
    "markdown>=3.5",
    "whitenoise>=6.7",
    "django-redis>=6.0.0",
    "psycopg2>=2.9.10",
]

[project.optional-dependencies]
postgres = [
    "psycopg[binary]>=3.2",
]
mysql = [
    "mysqlclient>=2.2",
]
prod = [
    "gunicorn>=21.2",
    "whitenoise>=6.7",
]
dev = [
    "ruff>=0.5.0",
    "black>=24.8.0",
    "isort>=5.13.2",
    "pytest>=8.2",
    "pytest-django>=4.9.0",
    "coverage>=7.6",
    "djlint>=1.35.2",
    "mypy>=1.10",
    "django-stubs>=5.0.0",
    "model-bakery>=1.19.5",
    "factory_boy>=3.3.0",
]

[project.urls]
Homepage = "https://github.com/yourname/echonote"
Repository = "https://github.com/yourname/echonote"
Issues = "https://github.com/yourname/echonote/issues"

[tool.black]
line-length = 100
target-version = ["py312"]
skip-string-normalization = false

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "DJ"]
ignore = []

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "echonote.settings"
pythonpath = ["."]
addopts = "-ra"
testpaths = [
    "tests",
]

[tool.coverage.run]
branch = true
source = ["echonote"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.django-stubs]
ignore_missing_model_attributes = true
django_settings_module = "echonote.settings"
